Changed the download script to stock download
