+++ docs/stylesheets/termynal.css
@@ -21,7 +21,7 @@
     /* font-family: 'Fira Mono', Consolas, Menlo, Monaco, 'Courier New', Courier, monospace; */
     font-family: 'Roboto Mono', 'Fira Mono', Consolas, Menlo, Monaco, 'Courier New', Courier, monospace;
     border-radius: 4px;
-    padding: 1px 45px 35px;
+    padding: 1px 45px 1px;
     position: relative;
     -webkit-box-sizing: border-box;
             box-sizing: border-box;
