+++ stockroom/download/__init__.py
@@ -1,0 +1,3 @@
+from stockroom.download.torchvision_datasets import gen_datasets
+from stockroom.download.pytorch_datasetloader import add_from_dataset
+__all__ = ['gen_datasets', 'add_from_dataset']
